+++ pytorch_forecasting/data/data_modules/__init__.py
@@ -1,0 +1,8 @@
+"""Data Modules (D2 Layer) of pytorch-forecasting v2"""
+
+from pytorch_forecasting.data.data_modules._encoder_decoder_data_module import (
+    EncoderDecoderTimeSeriesDataModule,
+)
+from pytorch_forecasting.data.data_modules._tslib_data_module import TslibDataModule
+
+__all__ = ["EncoderDecoderTimeSeriesDataModule", "TslibDataModule"]
