+++ pytorch_forecasting/models/samformer/_samformer_v2_pkg.py
@@ -23,7 +23,7 @@ class Samformer_pkg_v2(Base_pkg):
     @classmethod
     def get_datamodule_cls(cls):
         """Get the underlying DataModule class."""
-        from pytorch_forecasting.data.data_module import (
+        from pytorch_forecasting.data.data_modules import (
             EncoderDecoderTimeSeriesDataModule,
         )
 
@@ -41,7 +41,6 @@ class Samformer_pkg_v2(Base_pkg):
             `MyClass(**params)` or `MyClass(**params[i])` creates a valid test instance.
             `create_test_instance` uses the first (or only) dictionary in `params`
         """
-        import torch.nn as nn
 
         from pytorch_forecasting.metrics import QuantileLoss
