+++ pytorch_forecasting/models/temporal_fusion_transformer/_tft_pkg_v2.py
@@ -25,7 +25,7 @@ class TFT_pkg_v2(Base_pkg):
     @classmethod
     def get_datamodule_cls(cls):
         """Get the underlying DataModule class."""
-        from pytorch_forecasting.data.data_module import (
+        from pytorch_forecasting.data.data_modules import (
             EncoderDecoderTimeSeriesDataModule,
         )
