+++ pytorch_forecasting/models/timexer/_timexer_pkg_v2.py
@@ -27,7 +27,9 @@ class TimeXer_pkg_v2(Base_pkg):
     @classmethod
     def get_datamodule_cls(cls):
         """Get the underlying DataModule class."""
-        from pytorch_forecasting.data._tslib_data_module import TslibDataModule
+        from pytorch_forecasting.data.data_modules._tslib_data_module import (
+            TslibDataModule,
+        )
 
         return TslibDataModule
