+++ tests/test_models/test_tft_v2.py
@@ -4,7 +4,7 @@ import pytest
 import torch
 import torch.nn as nn
 
-from pytorch_forecasting.data.data_module import EncoderDecoderTimeSeriesDataModule
+from pytorch_forecasting.data.data_modules import EncoderDecoderTimeSeriesDataModule
 from pytorch_forecasting.data.timeseries import TimeSeries
 from pytorch_forecasting.models.temporal_fusion_transformer._tft_v2 import TFT
